Shift list command tests into separate class
